flask_sample/requirements.txt
attrs==22.1.0
beautifulsoup4==4.11.1
click==8.1.3
colorama==0.4.5
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
gunicorn==20.1.0
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
soupsieve==2.3.2.post1
tomli==2.0.1
virtualenv==20.16.5
Werkzeug==2.2.2
zipp==3.9.0
